Sending edition days ordered by date to the client
