--- resources/js/Components/Program/Day/NoInfo.vue
@@ -1,5 +1,0 @@
-<script setup lang="ts"></script>
-<template>
-    <p>Não conseguimos obter os resultados que pretendia.</p>
-    <p>Por favor, tente mais tarde.</p>
-</template>
